requirements.txt
Django==4.2
psycopg2-binary==2.9
openai==0.28
python-dotenv==1.0